feat(ui): 添加网格显示功能并调整实体尺寸
refactor(config): 更新世界尺寸和初始种群数量
调整世界尺寸为1920x1080，增加草和牛的初始数量以平衡生态系统

fix(simulation): 修正能量消耗类型为double
将energy_consumption从int改为double以支持更精确的能量计算

perf(ui): 提高最大缩放倍数至60
优化相机控制，允许用户更近距离观察模拟细节

chore: 禁用背景音乐并保持静音

+++ backend/include/animal.h
@@ -46,7 +46,7 @@ public:
     std::weak_ptr<Animal> mating_partner;
 
     double movement_speed;
-    int energy_consumption;
+    double energy_consumption;
     double hunting_range;
     double hunting_success_rate;
     double detection_range;
